Updated the readme.txt file to validate for WP SVN repository.
Added an "Upgrade" section, a donate link, and a "Screenshots" section with
2 screenshots, one a simple use case and another a use case that caused me
to create this plugin in the first place that is actually useful.

For the donate link, since it needed something in order to validate, I just
picked a Charity Navigator link to Camp Sunshine, an organization I have
raised money for before. If somebody finds this plugin useful, then they can
donate to their own favorite charity or use my suggestion. For something
this simple, I don't feel a need for a direct donation.